temporarily disable testing in browser
Currently, `request-master` uses `karma-phantomjs-launcher` and `phantomjs-prebuilt` to run tests in browser. The problem is that one of the dependencies uses `const` keyword but `phantomjs` doesn't know about it so it triggers error. Even worse, it's actually severely outdated because it uses `WebKit` underneath. We should upgrade to some headless browser anyway soon.